# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Labloom API",
    "version": "1.0.0",
    "description": "API documentation for Labloom Healthcare Application"
  },
  "servers": [
    {
      "url": "https://labloom-new.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/request-otp": {
      "post": {
        "summary": "Request OTP for admin login",
        "tags": [
          "Admin Portal"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent"
          }
        }
      }
    },
    "/api/admin/verify-otp": {
      "post": {
        "summary": "Verify OTP and login as admin",
        "tags": [
          "Admin Portal"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "otp"
                ],
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "otp": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
    "/api/admin/pending-hospitals": {
      "get": {
        "summary": "List hospitals awaiting verification",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending hospitals"
          }
        }
      }
    },
    "/api/admin/approve-hospital/{id}": {
      "post": {
        "summary": "Approve hospital registration",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hospital approved"
          }
        }
      }
    },
    "/api/admin/pending-labs": {
      "get": {
        "summary": "List labs awaiting verification",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending labs"
          }
        }
      }
    },
    "/api/admin/approve-lab/{id}": {
      "post": {
        "summary": "Approve lab registration",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lab approved"
          }
        }
      }
    },
    "/api/admin/pending-doctors": {
      "get": {
        "summary": "List doctors awaiting verification",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending doctors"
          }
        }
      }
    },
    "/api/admin/approve-doctor/{id}": {
      "post": {
        "summary": "Approve doctor registration",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor approved"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "Get all system users (excludes admins)",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      }
    },
    "/api/admin/users/{id}/status": {
      "patch": {
        "summary": "Update user account status (Active/Suspended)",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        }
      }
    },
    "/api/admin/reports/system": {
      "get": {
        "summary": "Get system-wide platform analytics",
        "tags": [
          "Admin Portal"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics data"
          }
        }
      }
    },
...
  },
...
}